cdk-3tier/requirements.txt
aws-cdk.core
aws.cdk.aws-s3
aws-cdk.aws-ec2
aws-cdk.aws-ssm
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws_autoscaling
aws-cdk.aws-rds